dnode service structure updated, config added, utils methods added
