--- testing/services/dummy/dmetrics/Dockerfile
@@ -1,5 +1,0 @@
-FROM alpine:3.13.5
-
-COPY bin/dmetric /usr/bin/dmetric
-
-CMD ["/usr/bin/dmetric"]
