--- testing/services/dummy/dmetrics/Int.Dockerfile
@@ -1,3 +1,0 @@
-FROM 003664043471.dkr.ecr.us-east-1.amazonaws.com/e2e-base-image:latest 
-
-ENV PROJECT_NAME=dummy-node
