+++ testing/services/dummy/dnode/Makefile
@@ -6,11 +6,9 @@
 
 include ../config.mk
 
-.PHONY: integration.test test
-
-build: integration.build
+build:
 	@echo Building version: \"$(BIN_VER)\"
-	env CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -ldflags='-X github.com/ukama/ukama/testing/services/dummy-node/dnode/cmd/version.Version=$(BIN_VER) -extldflags=-static' -o bin/dnode cmd/main.go
+	env CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -ldflags='-X github.com/ukama/ukama/testing/services/dummy/dnode/version.Version=$(BIN_VER) -extldflags=-static' -o bin/dnode server/main.go
 
 test:
 	go test -v ./...
@@ -23,19 +21,8 @@ deps:
 			google.golang.org/grpc/cmd/protoc-gen-go-grpc \
 			github.com/ysugimoto/grpc-graphql-gateway/protoc-gen-graphql/...
 			github.com/vektra/mockery/v2/.../
-
-gen:
-	protoc -I ./pb -I ../../../../systems/common/pb -I ../../../../systems/common/pb/dep --go_out=$(GOPATH)/src --go-grpc_out=$(GOPATH)/src --go-grpc_opt=require_unimplemented_servers=true --govalidators_out=$(GOPATH)/src \
- 			  dnode.proto
-	mockery --all --recursive --dir ./internal
-	mockery --all --recursive --dir ./pkg
-	mockery --dir ./pb  --all --recursive --output ./pb/gen/mocks
-
 clean:
 	rm -rf bin/
 
-integration.test:
-	go test ./test/integration -tags integration  -v -count=1
-
-integration.build:
-	env CGO_ENABLED=0 go test ./test/integration -tags integration -v -c -o bin/integration
+server:
+	go run server/main.go
