--- testing/services/dummy/dnode/cmd/main.go
@@ -1,103 +1,0 @@
-/*
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- *
- * Copyright (c) 2023-present, Ukama Inc.
- */
-
-package main
-
-import (
-	"os"
-
-	"github.com/num30/config"
-	"google.golang.org/grpc"
-	yaml "gopkg.in/yaml.v2"
-
-	log "github.com/sirupsen/logrus"
-	ccmd "github.com/ukama/ukama/systems/common/cmd"
-	ugrpc "github.com/ukama/ukama/systems/common/grpc"
-	"github.com/ukama/ukama/systems/common/msgBusServiceClient"
-	mb "github.com/ukama/ukama/systems/common/msgBusServiceClient"
-	"github.com/ukama/ukama/systems/common/uuid"
-	"github.com/ukama/ukama/testing/services/dummy-node/dnode/cmd/version"
-	generated "github.com/ukama/ukama/testing/services/dummy-node/dnode/pb/gen"
-	"github.com/ukama/ukama/testing/services/dummy-node/dnode/pkg"
-	"github.com/ukama/ukama/testing/services/dummy-node/dnode/pkg/server"
-)
-
-var serviceConfig *pkg.Config
-
-func main() {
-	ccmd.ProcessVersionArgument(pkg.ServiceName, os.Args, version.Version)
-	initConfig()
-	runGrpcServer()
-}
-func initConfig() {
-	serviceConfig = pkg.NewConfig(pkg.ServiceName)
-	err := config.NewConfReader(pkg.ServiceName).Read(serviceConfig)
-	if err != nil {
-		log.Fatal("Error reading config ", err)
-	} else if serviceConfig.DebugMode {
-		b, err := yaml.Marshal(serviceConfig)
-		if err != nil {
-			log.Infof("Config:\n%s", string(b))
-		}
-	}
-	pkg.IsDebugMode = serviceConfig.DebugMode
-}
-
-func runGrpcServer() {
-
-	instanceId := os.Getenv("POD_NAME")
-	if instanceId == "" {
-		inst := uuid.NewV4()
-		instanceId = inst.String()
-	}
-
-	mbClient := msgBusServiceClient.NewMsgBusClient(serviceConfig.MsgClient.Timeout,
-		serviceConfig.OrgName, pkg.SystemName, pkg.ServiceName, instanceId, serviceConfig.Queue.Uri,
-		serviceConfig.Service.Uri, serviceConfig.MsgClient.Host, serviceConfig.MsgClient.Exchange,
-		serviceConfig.MsgClient.ListenQueue, serviceConfig.MsgClient.PublishQueue,
-		serviceConfig.MsgClient.RetryCount, serviceConfig.MsgClient.ListenerRoutes)
-
-	nodeServer := server.NewNodeServer(serviceConfig.OrgName, mbClient, serviceConfig.AmqpConfig)
-
-	log.Debugf("MessageBus Client is %+v", mbClient)
-
-	grpcServer := ugrpc.NewGrpcServer(*serviceConfig.Grpc, func(s *grpc.Server) {
-		generated.RegisterNodeServiceServer(s, nodeServer)
-	})
-
-	go grpcServer.StartServer()
-
-	go msgBusListener(mbClient)
-
-	waitForExit()
-}
-
-func msgBusListener(m mb.MsgBusServiceClient) {
-
-	if err := m.Register(); err != nil {
-		log.Errorf("Failed to register to Message Client Service. Error %s", err.Error())
-	}
-
-	if err := m.Start(); err != nil {
-		log.Errorf("Failed to start to Message Client Service routine for service %s. Error %s", pkg.ServiceName, err.Error())
-	}
-}
-
-func waitForExit() {
-	sigs := make(chan os.Signal, 1)
-	done := make(chan bool, 1)
-	go func() {
-		sig := <-sigs
-		log.Info(sig)
-		done <- true
-	}()
-
-	log.Debug("awaiting terminate/interrrupt signal")
-	<-done
-	log.Infof("exiting service %s", pkg.ServiceName)
-}
