--- testing/services/dummy/dnode/pb/dnode.proto
@@ -1,29 +1,0 @@
-/*
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- *
- * Copyright (c) 2023-present, Ukama Inc.
- */
-
- syntax = "proto3";
-
- option go_package = "github.com/ukama/ukama/testing/services/dummy-node/dnode/pb/gen";
-  
- package ukama.dummynode.dnode.v1;
- 
- service NodeService { 
-    rpc ResetNode(Request) returns (Response) {}
-    rpc NodeRFOn(Request) returns (Response) {}
-    rpc NodeRFOff(Request) returns (Response) {}
-    rpc TurnNodeOff(Request) returns (Response) {}
-    rpc TurnNodeOnline(Request) returns (Response) {}
- }
- 
-message Request {
-    string nodeId = 1 [json_name = "node_id"];
-}
-
-message Response {
-    string nodeId = 1 [json_name = "node_id"];
-}
