--- testing/services/dummy/dnode/pkg/config.go
@@ -1,43 +1,0 @@
-/*
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- *
- * Copyright (c) 2023-present, Ukama Inc.
- */
-
-package pkg
-
-import (
-	"time"
-
-	uconf "github.com/ukama/ukama/systems/common/config"
-)
-
-type Config struct {
-	uconf.BaseConfig `mapstructure:",squash"`
-	Grpc             *uconf.Grpc      `default:"{}"`
-	Queue            *uconf.Queue     `default:"{}"`
-	Timeout          time.Duration    `default:"3s"`
-	MsgClient        *uconf.MsgClient `default:"{}"`
-	Service          *uconf.Service
-	OrgName          string
-	AmqpConfig       AmqpConfig
-}
-
-type AmqpConfig struct {
-	Uri      string `default:"http://rabbitmq:15672"`
-	Username string `default:"guest"`
-	Password string `default:"guest"`
-	Exchange string `default:"amq.topic"`
-	Vhost    string `default:"%2F"`
-}
-
-func NewConfig(name string) *Config {
-	return &Config{
-		Service: uconf.LoadServiceHostConfig(name),
-		MsgClient: &uconf.MsgClient{
-			Timeout: 5 * time.Second,
-		},
-	}
-}
