--- testing/services/dummy/dnode/pkg/global.go
@@ -1,14 +1,0 @@
-/*
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- *
- * Copyright (c) 2023-present, Ukama Inc.
- */
-
-package pkg
-
-const ServiceName = "dnode"
-const SystemName = "dummynode"
-
-var IsDebugMode = false
