--- testing/services/dummy/dnode/test/integration/integration_test.go
@@ -1,41 +1,0 @@
-//go:build integration
-// +build integration
-
-/*
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- *
- * Copyright (c) 2023-present, Ukama Inc.
- */
-
-package integration
-
-import (
-	log "github.com/sirupsen/logrus"
-	"github.com/ukama/ukama/systems/common/config"
-
-	rconf "github.com/num30/config"
-)
-
-var tConfig *TestConfig
-
-func init() {
-	// load config
-	tConfig = &TestConfig{}
-
-	reader := rconf.NewConfReader("integration")
-
-	err := reader.Read(tConfig)
-	if err != nil {
-		log.Errorf("Failed to read config: %v", err)
-	}
-
-	log.Info("Expected config ", "integration.yaml", " or env vars for ex: SERVICEHOST")
-	log.Infof("Config: %+v\n", tConfig)
-}
-
-type TestConfig struct {
-	ServiceHost string        `default:"localhost:9090"`
-	Queue       *config.Queue `default:"{}"`
-}
