+++ testing/services/dummy/dnode/utils/pushEvents.go
@@ -19,7 +19,7 @@ import (
 	mb "github.com/ukama/ukama/systems/common/msgBusServiceClient"
 	"github.com/ukama/ukama/systems/common/msgbus"
 	epb "github.com/ukama/ukama/systems/common/pb/gen/events"
-	"github.com/ukama/ukama/testing/services/dummy-node/dnode/pkg"
+	"github.com/ukama/ukama/testing/services/dummy/dnode/config"
 	"google.golang.org/protobuf/types/known/anypb"
 )
 
@@ -31,7 +31,7 @@ func getRoutingKey(orgName string) msgbus.RoutingKeyBuilder {
 	return msgbus.NewRoutingKeyBuilder().SetCloudSource().SetSystem("messaging").SetOrgName(orgName).SetService("mesh")
 }
 
-func Run(amqpConf pkg.AmqpConfig, route string, payload *anypb.Any, out io.Writer) error {
+func Run(amqpConf config.AmqpConfig, route string, payload *anypb.Any, out io.Writer) error {
 
 	aClient := NewAmqpClient(amqpConf, defaultDuration)
 
@@ -70,7 +70,7 @@ func PushNodeOnline(orgName, nodeId string, m mb.MsgBusServiceClient) {
 	}
 }
 
-func PushNodeOnlineViaREST(amqpConf pkg.AmqpConfig, org, nodeId string, m mb.MsgBusServiceClient) {
+func PushNodeOnlineViaREST(amqpConf config.AmqpConfig, org, nodeId string, m mb.MsgBusServiceClient) {
 	route := getRoutingKey(org).SetAction("online").SetObject("node").MustBuild()
 
 	msg := epb.NodeOnlineEvent{
@@ -106,7 +106,7 @@ func PushNodeReset(orgName, nodeId string, m mb.MsgBusServiceClient) {
 	}
 }
 
-func PushNodeResetViaREST(amqpConf pkg.AmqpConfig, org, nodeId string, m mb.MsgBusServiceClient) {
+func PushNodeResetViaREST(amqpConf config.AmqpConfig, org, nodeId string, m mb.MsgBusServiceClient) {
 	route := getRoutingKey(org).SetAction("reset").SetObject("node").MustBuild()
 
 	msg := epb.NodeOnlineEvent{
@@ -142,7 +142,7 @@ func PushNodeRFOn(orgName, nodeId string, m mb.MsgBusServiceClient) {
 	}
 }
 
-func PushNodeRFOnViaREST(amqpConf pkg.AmqpConfig, org, nodeId string, m mb.MsgBusServiceClient) {
+func PushNodeRFOnViaREST(amqpConf config.AmqpConfig, org, nodeId string, m mb.MsgBusServiceClient) {
 	route := getRoutingKey(org).SetAction("rfon").SetObject("node").MustBuild()
 
 	msg := epb.NodeOnlineEvent{
@@ -178,7 +178,7 @@ func PushNodeRFOff(orgName, nodeId string, m mb.MsgBusServiceClient) {
 	}
 }
 
-func PushNodeRFOffViaREST(amqpConf pkg.AmqpConfig, org, nodeId string, m mb.MsgBusServiceClient) {
+func PushNodeRFOffViaREST(amqpConf config.AmqpConfig, org, nodeId string, m mb.MsgBusServiceClient) {
 	route := getRoutingKey(org).SetAction("rfoff").SetObject("node").MustBuild()
 
 	msg := epb.NodeOnlineEvent{
@@ -211,7 +211,7 @@ func PushNodeOff(orgName, nodeId string, m mb.MsgBusServiceClient) {
 	}
 }
 
-func PushNodeOffViaREST(amqpConf pkg.AmqpConfig, org, nodeId string, m mb.MsgBusServiceClient) {
+func PushNodeOffViaREST(amqpConf config.AmqpConfig, org, nodeId string, m mb.MsgBusServiceClient) {
 	route := getRoutingKey(org).SetAction("off").SetObject("node").MustBuild()
 
 	msg := epb.NodeOfflineEvent{
