+++ testing/services/dummy/dnode/version/version.go
@@ -9,4 +9,5 @@
 package version
 
 // Minor version is autoupdated by the build system
+// NOTE: use go build -ldflags "-X github.com/ukama/ukama/systems/registry/network/cmd/version.Version==$(git describe)".
 var Version = "v0.0.debug"
